Fix for issue #4
If an entry has more than 20 authors, the function `apa_names(...)` in
`formatting/apa.py` switches to a block of code which has a minor bug on
line 40: it references `richtext.Text` but `richtext` is not imported –
what is imported instead is `Text` without the module name.

Solution: replace `richtext.Text` with `Text`.